fix(argcount) fix n if too low
With Luassert, an assertion is responsible for passing the args
to 'say'. The args table is often passed along from the internal
luassert one to 'say'. If, for some messages, an extra argument is
inserted, and the assertion does not increase the existing 'n' field
on that table, then 'say' will use 'n' and be one argument short.